refactor: WEBAPP 폴더 구조 개편 및 확장성 개선
- datas 폴더를 api와 constants로 분리
- pages 폴더 생성 및 페이지 컴포넌트 분리
- config 폴더를 통한 환경 설정 관리
- README 업데이트 및 프로젝트 문서화 완성
- 모든 import 경로 최신화
